+++ src/c/breseq/libbreseq/chisquare.h
@@ -9,9 +9,9 @@ namespace breseq {
 	double incompletegamma(double a, double x, bool complemented = false);
 	double lngamma(double x, double* sgngam);
   
-  double nbdtrc(int k, double n, double p);
-  double nbdtr(int k, double n, double p);
-  double nbdtri(int k, double n, double p);
+  double nbdtrc(double k, double n, double p);
+  double nbdtr(double k, double n, double p);
+  double nbdtri(double k, double n, double p);
   
   double incbi(double aa, double bb, double yy0);
   double incbet(double aa, double bb, double xx);
@@ -21,8 +21,8 @@ namespace breseq {
   
   double ndtri(double y0);
   
-  double bdtrc(int k, double n, double p);
-  double bdtr(int k, double n, double p);
-  double bdtri(int k, double n, double y);
+  double bdtrc(double k, double n, double p);
+  double bdtr(double k, double n, double p);
+  double bdtri(double k, double n, double y);
 
 } // namespace breseq
